# docs/atlas-prd/03-contracts/openapi.yaml
openapi: 3.1.1
info:
  title: Atlas Wallet and Financial Operations API
  version: 2026-07-23
  summary: Security-first contracts for wallet, transfers, merchant payments, operations, and reporting.
  description: |
    Reference contract for the Atlas portfolio platform. Financial amounts are integer minor-unit
    strings. Money-moving commands are idempotent. Browser applications use a backend-for-frontend
    session; merchant and service integrations use scoped machine credentials. This file is a
    deliberately curated baseline, not an automatically inferred description of handlers.
  contact:
    name: Atlas API Governance
  license:
    name: Portfolio reference - no production authorization
servers:
  - url: https://api.staging.atlas.example
    description: Synthetic staging environment
  - url: http://localhost:8080
    description: Local development
...
security:
  - bffSession: []
paths:
...
  /v1/approvals:
    get:
      operationId: listApprovals
      tags: [Approvals]
      summary: List approvals visible to the active tenant and principal
      parameters:
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Approval page
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [data, page]
                properties:
                  data:
                    type: array
                    items: {$ref: '#/components/schemas/Approval'}
                  page: {$ref: '#/components/schemas/PageInfo'}
        '400': {$ref: '#/components/responses/RequestMalformed'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
    post:
      operationId: createApproval
      tags: [Approvals]
      summary: Create a typed approval request over a canonical payload digest
      parameters:
        - $ref: '#/components/parameters/XAtlasCSRFToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateApprovalRequest'}
      responses:
        '201':
          description: Approval request created
          headers:
            ETag: {$ref: '#/components/headers/ETag'}
            Idempotency-Replayed: {$ref: '#/components/headers/IdempotencyReplayed'}
            Location: {$ref: '#/components/headers/Location'}
            X-Authorization-Decision-Id: {$ref: '#/components/headers/XAuthorizationDecisionId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Approval'}
        '400': {$ref: '#/components/responses/RequestMalformed'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/StepUpRequired'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/ValidationFailed'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /v1/approvals/{approval_id}:
    parameters:
      - $ref: '#/components/parameters/ApprovalId'
    get:
      operationId: getApproval
      tags: [Approvals]
      summary: Get an approval and its immutable payload digest
      responses:
        '200':
          description: Approval detail
          headers:
            ETag: {$ref: '#/components/headers/ETag'}
            X-Authorization-Decision-Id: {$ref: '#/components/headers/XAuthorizationDecisionId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Approval'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFoundOrConcealed'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /v1/approvals/{approval_id}/decisions:
    parameters:
      - $ref: '#/components/parameters/ApprovalId'
    post:
      operationId: decideApproval
      tags: [Approvals]
      summary: Approve or reject an approval as a distinct principal
      parameters:
        - $ref: '#/components/parameters/XAtlasCSRFToken'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ApprovalDecisionRequest'}
      responses:
        '200':
          description: Approval decision recorded
          headers:
            ETag: {$ref: '#/components/headers/ETag'}
            Idempotency-Replayed: {$ref: '#/components/headers/IdempotencyReplayed'}
            Location: {$ref: '#/components/headers/Location'}
            X-Authorization-Decision-Id: {$ref: '#/components/headers/XAuthorizationDecisionId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Approval'}
        '400': {$ref: '#/components/responses/RequestMalformed'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/StepUpRequired'}
        '404': {$ref: '#/components/responses/NotFoundOrConcealed'}
        '409': {$ref: '#/components/responses/Conflict'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /v1/approvals/{approval_id}/executions:
    parameters:
      - $ref: '#/components/parameters/ApprovalId'
    post:
      operationId: executeApproval
      tags: [Approvals]
      summary: Execute an approved typed action after payload revalidation
      parameters:
        - $ref: '#/components/parameters/XAtlasCSRFToken'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '200':
          description: Approved action executed
          headers:
            ETag: {$ref: '#/components/headers/ETag'}
            Idempotency-Replayed: {$ref: '#/components/headers/IdempotencyReplayed'}
            Location: {$ref: '#/components/headers/Location'}
            X-Authorization-Decision-Id: {$ref: '#/components/headers/XAuthorizationDecisionId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Approval'}
        '400': {$ref: '#/components/responses/RequestMalformed'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/StepUpRequired'}
        '404': {$ref: '#/components/responses/NotFoundOrConcealed'}
        '409': {$ref: '#/components/responses/Conflict'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
  /v1/approvals/{approval_id}/cancellations:
    parameters:
      - $ref: '#/components/parameters/ApprovalId'
    post:
      operationId: cancelApproval
      tags: [Approvals]
      summary: Cancel a pending approval
      parameters:
        - $ref: '#/components/parameters/XAtlasCSRFToken'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ApprovalCancellationRequest'}
      responses:
        '200':
          description: Approval cancelled
          headers:
            ETag: {$ref: '#/components/headers/ETag'}
            Idempotency-Replayed: {$ref: '#/components/headers/IdempotencyReplayed'}
            Location: {$ref: '#/components/headers/Location'}
            X-Authorization-Decision-Id: {$ref: '#/components/headers/XAuthorizationDecisionId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Approval'}
        '400': {$ref: '#/components/responses/RequestMalformed'}
        '401': {$ref: '#/components/responses/AuthenticationRequired'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFoundOrConcealed'}
        '409': {$ref: '#/components/responses/Conflict'}
        '412': {$ref: '#/components/responses/PreconditionFailed'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
...
components:
  securitySchemes:
    bffSession:
      type: apiKey
      in: cookie
      name: __Host-atlas_session
      description: Secure, HttpOnly, SameSite session cookie managed by the BFF.
    merchantApiKey:
      type: apiKey
      in: header
      name: Authorization
      description: |
        AtlasKey credential formatted as `AtlasKey <key-id>.<secret>`. Secrets are high-entropy,
        one-time-disclosed values stored only as SHA-256 verifiers. The `identity:read` scope
        authorizes only `GET /v1/me` in Phase 01.
    merchantOAuth:
      type: oauth2
      description: OAuth 2.0 client-credentials access token with audience and scope enforcement.
      flows:
        clientCredentials:
          tokenUrl: https://identity.staging.atlas.example/oauth2/token
          scopes:
            payments:read: Read payment intents
            payments:write: Create and capture payment intents
            refunds:read: Read refunds
            refunds:write: Create refunds
            webhooks:read: Read webhook endpoints and deliveries
            webhooks:write: Manage endpoints and replay deliveries
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Unique within authenticated actor, tenant, operation, and API version.
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: '^[A-Za-z0-9._:-]+$'
...
    XAtlasCSRFToken:
      name: X-Atlas-CSRF-Token
      in: header
      required: true
      description: Memory-only token bound to the current browser session and its latest rotation.
      schema:
        type: string
        minLength: 43
        maxLength: 128
        pattern: '^[A-Za-z0-9_-]+$'
...
    PageSize:
      name: page_size
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 25}
    Cursor:
      name: cursor
      in: query
      schema: {type: string, minLength: 16, maxLength: 2048}
...
    ApprovalId:
      name: approval_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/OpaqueId'}
...
  headers:
...
    ETag:
      schema: {type: string}
    IdempotencyReplayed:
      schema: {type: boolean}
      description: True when returning the persisted response for a previous identical request.
...
    XAuthorizationDecisionId:
      description: Opaque identifier linking a privileged or denied action to its authorization evidence.
      schema: {$ref: '#/components/schemas/OpaqueId'}
...
    Location:
      schema: {type: string, format: uri-reference, maxLength: 2048}
  responses:
    AuthenticationRequired:
      description: Authentication required or invalid
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
          example:
            type: https://atlas.example/problems/authentication-required
            title: Authentication required
            status: 401
            code: AUTHENTICATION_REQUIRED
            request_id: req_01JAT1AS00000000000001
            retryable: false
    StepUpRequired:
      description: A fresh higher-assurance authentication is required
      headers:
        X-Authorization-Decision-Id: {$ref: '#/components/headers/XAuthorizationDecisionId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
          example:
            type: https://atlas.example/problems/step-up-required
            title: Additional verification required
            status: 403
            code: STEP_UP_REQUIRED
            request_id: req_01JAT1AS00000000000001
            retryable: false
    Forbidden:
      description: Action is not authorized
      headers:
        X-Authorization-Decision-Id: {$ref: '#/components/headers/XAuthorizationDecisionId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
    RequestMalformed:
      description: Request method, query, or body shape is not supported by this operation
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
        X-Correlation-Id: {$ref: '#/components/headers/XCorrelationId'}
        traceparent: {$ref: '#/components/headers/Traceparent'}
        Cache-Control: {$ref: '#/components/headers/CacheControl'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
    ServiceUnavailable:
      description: Required dependency or migration state is not ready; no topology is disclosed
      headers:
        X-Request-Id: {$ref: '#/components/headers/XRequestId'}
        X-Correlation-Id: {$ref: '#/components/headers/XCorrelationId'}
        traceparent: {$ref: '#/components/headers/Traceparent'}
        Cache-Control: {$ref: '#/components/headers/CacheControl'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
          example:
            type: https://atlas.example/problems/dependency-degraded
            title: Service unavailable
            status: 503
            code: DEPENDENCY_DEGRADED
            request_id: req_01JAT1AS00000000000001
            correlation_id: cor_01JAT1AS00000000000001
            retryable: true
  schemas:
...
    CreateApprovalRequest:
      type: object
      additionalProperties: false
      required: [action_type, payload]
      properties:
        action_type:
          type: string
          const: identity.organization.membership.change_admin
        payload: {$ref: '#/components/schemas/MembershipRoleChangeApprovalRequest'}
    Approval:
      type: object
      additionalProperties: false
      required:
        [id, organization_id, action_type, payload_digest, status, requester_principal_id,
         purpose, version, expires_at, created_at]
      properties:
        id: {$ref: '#/components/schemas/OpaqueId'}
        organization_id: {$ref: '#/components/schemas/OpaqueId'}
        action_type:
          type: string
          enum: [identity.organization.membership.change_admin]
        payload_digest:
          type: string
          pattern: '^[a-f0-9]{64}$'
          description: SHA-256 over the RFC 8785 canonical request payload.
        status: {$ref: '#/components/schemas/ApprovalStatus'}
        requester_principal_id: {$ref: '#/components/schemas/OpaqueId'}
        decider_principal_id:
          oneOf: [{$ref: '#/components/schemas/OpaqueId'}, {type: 'null'}]
        purpose: {$ref: '#/components/schemas/Purpose'}
        decision_reason: {type: [string, 'null'], maxLength: 500}
        version: {type: integer, minimum: 1}
        expires_at: {type: string, format: date-time}
        created_at: {type: string, format: date-time}
        decided_at: {type: [string, 'null'], format: date-time}
        executed_at: {type: [string, 'null'], format: date-time}
...
    PageInfo:
      type: object
      additionalProperties: false
      required: [has_more]
      properties:
        next_cursor: {type: [string, 'null']}
        has_more: {type: boolean}
...